feat(engine): esbuild bundling + move original code above webworker code
