+++ package.json
@@ -9,7 +9,9 @@
   },
   "scripts": {
     "build": "npm run ts-loader src/lostcity/tools/cache/pack.ts",
-    "bundle": "bun Bundle.ts",
+    "bundle": "npm run bundle:esbuild",
+    "bundle:bun": "bun bundle.ts",
+    "bundle:esbuild": "npm run ts-loader bundle.ts",
     "clean": "npm run client:clean && npm run server:clean",
     "client": "java -cp 225.2004-05-18.jar client 10 0 highmem members",
     "client:clean": "npm run ts-loader src/lostcity/tools/cache/cleanClient.ts",
@@ -66,6 +68,7 @@
     "@types/node-forge": "^1.3.11",
     "@types/ws": "^8.5.10",
     "@typescript-eslint/eslint-plugin": "^7.9.0",
+    "esbuild": "^0.23.0",
     "eslint": "^8.57.0",
     "eslint-plugin-unused-imports": "^3.2.0",
     "eslint-plugin-vitest": "^0.5.4",
