+++ src/lostcity/util/WorkerFactory.ts
@@ -1,9 +1,9 @@
-import { Worker as NodeWorker } from 'node:worker_threads';
+import { Worker as NodeWorker } from 'worker_threads';
 
 export function createWorker(fileName: string): Worker | NodeWorker {
-    if (typeof self !== 'undefined') {
-        return new Worker(fileName, {type: 'module'});
-    } else {
+    if (typeof self === 'undefined') {
         return new NodeWorker(fileName);
+    } else {
+        return new Worker(fileName, {type: 'module'});
     }
 }
